Add CONTRIBUTING.md and CONTRIBUTORS.md files
Added a CONTRIBUTING.md which lays out guidelines and some helpful
information to help people who are new to contributing to the project.
This also has a slight amendment to the current CoC to help it be
relevant in this context as well, as current wording feels more geared
to events, rather than contributors to an OSS/F(L)OSS project.

Added a CONTRIBUTORS.md file which doubles as an contributor attribution
list, and as a sign-off that they agree to CC BY-SA 4.0 International
license.

Signed-off-by: Adam Jimerson <vendion@gmail.com>